Center of Gravity is calculated correctly now

+++ FuzzyProject/main.cpp
@@ -10,7 +10,7 @@ int main() {
 	int a, b, c;
 	while (std::cin >> a >> b >> c) {
 		
-		int dd = sf.getLeftOutput(a, b, c);
+		float dd = sf.getLeftOutput(a, b, c);
 		std::cout << "\n LEFT SPEED:" << dd << "\n";
 		dd = sf.getRightOutput(a, b, c);
 		std::cout << "RIGHT SPEED:" << dd << "\n";
